Inline connected users count within Google_Proxy.
